requirements.txt
pytest-playwright
pytest-asyncio
allure-pytest
playwright